@@ -1,56 +1,39 @@
 {
   "name": "ngx-pinch-zoom",
   "version": "20.0.0",
+  "description": "Angular pinch-zoom component with signals support for touch-enabled image zooming and panning",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "build-lib": "npx ng-packagr -p projects/ngx-pinch-zoom/ng-package.json",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e",
-    "cypress:open": "cypress open",
-    "cypress:run": "cypress run"
+    "build-lib": "ng build ngx-pinch-zoom",
+    "format": "prettier --write \"**/*.{ts,html,scss,json,md}\"",
+    "format:check": "prettier --check \"**/*.{ts,html,scss,json,md}\""
   },
   "dependencies": {
-    "@angular/animations": "^19.0.0",
-    "@angular/common": "^19.0.0",
-    "@angular/compiler": "^19.0.0",
-    "@angular/core": "^19.0.0",
-    "@angular/forms": "^19.0.0",
-    "@angular/platform-browser": "^19.0.0",
-    "@angular/platform-browser-dynamic": "^19.0.0",
-    "@angular/router": "^19.0.0",
-    "rxjs": "~7.5.2",
-    "tslib": "^2.3.1",
-    "zone.js": "0.15.0"
+    "@angular/animations": "^20.0.0",
+    "@angular/common": "^20.0.0",
+    "@angular/compiler": "^20.0.0",
+    "@angular/core": "^20.0.0",
+    "@angular/forms": "^20.0.0",
+    "@angular/platform-browser": "^20.0.0",
+    "@angular/platform-browser-dynamic": "^20.0.0",
+    "@angular/router": "^20.0.0",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.8.0",
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.0.4",
-    "@angular-eslint/eslint-plugin": "^19.0.0",
-    "@angular-eslint/eslint-plugin-template": "^19.0.0",
-    "@angular/cli": "^19.0.0",
-    "@angular/compiler-cli": "^19.0.0",
-    "@cypress/schematic": "^1.6.0",
-    "@types/jasmine": "~3.10.3",
-    "@types/node": "^16.18.24",
-    "@typescript-eslint/eslint-plugin": "^5.10.2",
-    "@typescript-eslint/parser": "^5.10.2",
-    "cypress": "latest",
-    "eslint": "^8.8.0",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jsdoc": "^37.7.1",
-    "eslint-plugin-prefer-arrow": "^1.2.3",
-    "jasmine-core": "~4.0.0",
-    "jasmine-spec-reporter": "~7.0.0",
-    "karma": "~6.3.15",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.1.1",
-    "karma-jasmine": "~4.0.1",
-    "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-packagr": "^19.0.0",
-    "prettier": "^2.7.1",
-    "ts-node": "~10.4.0",
-    "typescript": "5.5.4"
+    "@angular-devkit/build-angular": "^20.0.0",
+    "@angular/cli": "^20.0.0",
+    "@angular/compiler-cli": "^20.0.0",
+    "@types/node": "^22.0.0",
+    "ng-packagr": "^20.0.0",
+    "prettier": "^3.0.0",
+    "typescript": "~5.8.0"
+  },
+  "engines": {
+    "node": "^18.19.1 || ^20.11.1 || ^22.0.0",
+    "npm": "^9.0.0 || ^10.0.0"
   }
 }
